dogexercise/requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12,<4.0
gunicorn>=20.1.0
psycopg2>=2.9.1

